More chasing that float64...

+++ Embeddings/positional_embedding.py
@@ -14,7 +14,7 @@ class positional_embedding(object):
 
         self.embeddings = cp.random.normal(0, .02, size=(self.max_seq_len, input_layer_shape)).astype(cp.float32)
 
-        self.dL_dE = cp.zeros_like(self.embeddings)
+        self.dL_dE = cp.zeros_like(self.embeddings).astype(cp.float32)
 
 ####################################
 # Forward Pass #
